+++ packages/vscode-extension/src/commands/convert.ts
@@ -1,26 +1,24 @@
-import { commands, Range, window } from "vscode";
-import { latexToPretext } from "../latextopretext";
+import { Range, window } from "vscode";
 import { markdownToPretext } from "md2ptx";
 import { pretextOutputChannel } from "../ui";
 import { convertToPretext } from "../importFiles";
 import { processLatexViaUnified } from "@unified-latex/unified-latex";
 import { htmlLike } from "@unified-latex/unified-latex-util-html-like";
-import { getArgsContent } from "@unified-latex/unified-latex-util-arguments";
 import {
   unifiedLatexToPretext,
-  unifiedLatexWrapPars,
   xmlCompilePlugin,
 } from "@unified-latex/unified-latex-to-pretext";
-import { formatPTX } from "../formatter";
 // @ts-ignore
 import { FlexTeXtConvert } from "frankenmarkup";
-import { get } from "http";
 import { Environment, Macro } from "@unified-latex/unified-latex-types";
 import { lspFormatText } from "../lsp-client/main";
+import { fromXml } from "xast-util-from-xml";
+import { toXml } from "xast-util-to-xml";
+import { SKIP, visit } from "unist-util-visit";
+import { fromMarkdown } from "mdast-util-from-markdown";
 
-export function cmdConvertToPretext() {
-  console.log("Converting to PreTeXt");
-  pretextOutputChannel.append("Converting to PreTeXt");
+export function cmdConvertFile() {
+  pretextOutputChannel.append("Converting selected file to PreTeXt");
   // show quick pick to select whether to convert with pandoc or plastex
   window
     .showQuickPick(["plastex", "pandoc"], {
@@ -35,41 +33,64 @@ export function cmdConvertToPretext() {
     });
 }
 
-export function cmdLatexToPretext() {
+export async function cmdConvertText() {
   const editor = window.activeTextEditor;
-  if (editor) {
-    const selection = editor.selection;
-    const selectionRange = selection.isEmpty
-      ? editor.document.lineAt(selection.start.line).range
-      : new Range(selection.start, selection.end);
-
-    //console.log("selectionRange is", selectionRange);
-
-    const initialText = editor.document.getText(selectionRange);
-
-    //var newText = latexToPretext(initialText);
-    let newText = convertWithUnified(initialText);
-
-    // Remove the starting <p> tag if we selected text in the middle of a line.
-    const pTagMatch = newText.match(/^<p>/);
-    if (pTagMatch && selectionRange.start.character > 0) {
-      newText = newText.replace(/^<p>/, "");
-    }
-
-    // Split consecutive tags with a space if present before formatting.
-    const formattedNewText = formatPTX(newText.replace(/(>)(<)/g, "$1 $2"));
-
-    //The prettier formatter is finicky; needs correct structure to work?
-
-    //const formattedNewText = prettier.format(newText, {
-    //  parser: "ptx",
-    //  plugins: [prettierPluginPretext as unknown as prettier.Plugin],
-    //});
-
-    editor.edit((editbuilder) => {
-      editbuilder.replace(selectionRange, formattedNewText);
-    });
+  if (!editor) {
+    pretextOutputChannel.appendLine("No active editor found to convert text.");
+    return;
   }
+  const selection = editor.selection;
+  const selectionRange = selection.isEmpty
+    ? editor.document.lineAt(selection.start.line).range
+    : new Range(selection.start, selection.end);
+
+  const initialText = editor.document.getText(selectionRange);
+  let convertedText: string;
+
+  pretextOutputChannel.appendLine(
+    "Converting selected text to PreTeXt format."
+  );
+  window
+    .showQuickPick(["LaTeX", "Markdown", "Mixed text"], {
+      placeHolder: "Which format is the selected text?",
+    })
+    .then(async (qpSelection) => {
+      if (!qpSelection) {
+        return;
+      }
+      switch (qpSelection) {
+        case "LaTeX":
+          convertedText = await cmdLatexToPretext(initialText, selectionRange);
+          break;
+        case "Markdown":
+          convertedText = await markdownToPretext(initialText);
+          break;
+        case "Mixed text":
+          convertedText = await cmdConvertMixedtextToPretext(initialText);
+          break;
+      }
+    })
+    .then(() => {
+      if (convertedText) {
+        editor.edit((editBuilder) => {
+          editBuilder.replace(selectionRange, convertedText);
+        });
+      }
+    });
+}
+
+async function cmdLatexToPretext(initialText: string, selectionRange: Range) {
+  //var newText = latexToPretext(initialText);
+  let newText = convertWithUnified(initialText);
+
+  // Remove the starting <p> tag if we selected text in the middle of a line.
+  const pTagMatch = newText.match(/^<p>/);
+  if (pTagMatch && selectionRange.start.character > 0) {
+    newText = newText.replace(/^<p>/, "");
+  }
+
+  // Split consecutive tags with a space if present before formatting.
+  return lspFormatText(newText.replace(/(>)(<)/g, "$1 $2"));
 }
 
 //const myMacroReplacements = {
@@ -125,53 +146,122 @@ function convertWithUnified(text: string) {
   return convert(text).value as string;
 }
 
-export function cmdMarkdownToPretext() {
-  // Give warning that this function is currently unvailable.
-  pretextOutputChannel.appendLine(
-    "Markdown to PreTeXt conversion is still very experiemental.  Use with care."
-  );
-
-  const editor = window.activeTextEditor;
-  if (editor) {
-    const selection = editor.selection;
-    const selectionRange = selection.isEmpty
-      ? editor.document.lineAt(selection.start.line).range
-      : new Range(selection.start, selection.end);
-
-    const initialText = editor.document.getText(selectionRange);
-
-    var newText = markdownToPretext(initialText);
-
-    editor.edit((editbuilder) => {
-      editbuilder.replace(selectionRange, newText);
-    });
-  }
-}
-
-export async function cmdConvertMixedtextToPretext() {
+async function cmdConvertMixedtextToPretext(initialText: string) {
   pretextOutputChannel.appendLine(
     "Mixed text to PreTeXt conversion is still very experiemental.  Use with care."
   );
+  const newText = FlexTeXtConvert(initialText);
+  return lspFormatText(newText);
+}
+
+////////////////// Experiments /////////////////////
+
+export async function cmdExperimentConvert() {
   const editor = window.activeTextEditor;
-  if (editor) {
-    const selection = editor.selection;
-    let selectionRange: Range;
-    let fullDocument = false;
-    if (selection.isEmpty) {
-      fullDocument = true;
-      selectionRange = new Range(
-        editor.document.positionAt(0),
-        editor.document.positionAt(editor.document.getText().length)
-      );
-    } else {
-      selectionRange = new Range(selection.start, selection.end);
-    }
-    const initialText = editor.document.getText(selectionRange);
-    const newText = FlexTeXtConvert(initialText);
-    const formattedNewText = await lspFormatText(newText);
-    editor.edit((editbuilder) => {
-      editbuilder.replace(selectionRange, formattedNewText);
-    });
-    pretextOutputChannel.appendLine("Mixed text converted to PreTeXt.");
+  if (!editor) {
+    pretextOutputChannel.appendLine("No active editor found to convert text.");
+    return;
   }
+  const selection = editor.selection;
+  const selectionRange = selection.isEmpty
+    ? editor.document.lineAt(selection.start.line).range
+    : new Range(selection.start, selection.end);
+
+  const initialText = editor.document.getText(selectionRange);
+  let convertedText: string;
+
+  // Prompt user to select a conversion method
+  pretextOutputChannel.appendLine(
+    "Experimental conversion functions are designed for testing and may not work as expected."
+  );
+  window
+    .showQuickPick([
+      {
+        label: "Use mdast",
+        description: "Preprocess selected text with Markdown AST",
+        function: "useMdast",
+      },
+      {
+        label: "Use xast",
+        description: "Convert selected text to PreTeXt format",
+        function: "useXast",
+      },
+    ])
+    .then(async (selection) => {
+      if (!selection) {
+        return;
+      }
+      switch (selection.function) {
+        case "useMdast":
+          convertedText = await convertPmdWithMdast(initialText);
+          break;
+        case "useXast":
+          convertedText = await convertPmdWithXast(initialText);
+          break;
+      }
+    })
+    .then(() => {
+      if (convertedText) {
+        editor.edit((editbuilder) => {
+          editbuilder.replace(selectionRange, convertedText);
+        });
+      }
+    });
+}
+
+async function convertPmdWithMdast(initialText: string) {
+  // Remove leading and trailing whitespace
+  const trimmedText = initialText.trim();
+  console.log("initialText is", trimmedText);
+
+  // Converte to mdast:
+  const tree = fromMarkdown(trimmedText);
+
+  console.log("mdast before: ", tree);
+
+  //const convertedText = toXml(tree);
+
+  // Convert the text using FlexTeXtConvert
+  //const convertedText = FlexTeXtConvert(trimmedText);
+  const convertedText = initialText;
+  console.log("converted text is", convertedText);
+
+  // Format the converted text
+  return lspFormatText(convertedText);
+}
+
+async function convertPmdWithXast(initialText: string) {
+  console.log("initialText is", initialText);
+
+  const tree = fromXml(`<root>${initialText}</root>`);
+
+  console.log("xast before: ", tree);
+
+  visit(tree, (node, index, parent) => {
+    if (node.type === "text") {
+      const converted = FlexTeXtConvert(node.value);
+
+      console.log("converted text is", converted);
+
+      const subtree = fromXml(converted);
+      // replace the node with the subtree
+      if (typeof index !== "number" || !parent) return;
+
+      parent.children.splice(index, 1, ...subtree.children);
+      return SKIP;
+    }
+  });
+
+  console.log("xast after: ", tree);
+  // Convert the resulting tree back to XML
+  let newXml = toXml(tree);
+
+  console.log("back to xml: ", newXml);
+  // strip the <root> and </root> tags
+  const rootTagMatch = newXml.match(/^<root>\s*([\s\S]*?)\s*<\/root>$/);
+  if (rootTagMatch) {
+    newXml = rootTagMatch[1];
+  }
+
+  return lspFormatText(newXml);
 }