+++ packages/vscode-extension/src/ui.ts
@@ -3,7 +3,6 @@ import {
   StatusBarAlignment,
   StatusBarItem,
   Terminal,
-  WebviewPanel,
   window,
 } from "vscode";
 import { updateStatusBarItem } from "./utils";